+++ templates/index.html
@@ -52,7 +52,7 @@
     </section>
     <div class="attractions_outter">
       <div class="attractions_group">
-        <!-- <h2 class="error_message">123</h2> -->
+        <!-- <h2 class="error_message">Error</h2> -->
         <!-- <div class="attraction">
           <div class="attraction_img">
             <div class="attraction_name_group">
@@ -72,6 +72,6 @@
     </footer>
   </div>
 
-  <script src="/js/index.js"></script>
+  <script type="module" src="/js/index.js"></script>
 </body>
 </html>
